feat(SPV-1249): logic moved to jsonrequire package
